First pass at OCF resource agent for LizardFS
This resource agent will manage LizardFS master and metalogger instances to
provide high-availability failover.  It does so by implementing an OCF
resource agent which is designed to be used as a master/slave resource.  An
OCF master/slave resource initially starts in the slave state, which in our
case means a metalogger, and is later promoted to a master by running an
mfsmetarestore operation.

When initially started as metalogger, it doesn't know the address of the
master, this is communicated in a later "notify" event.  So we start the
metalogger with an invalid master and change the configuration and reload
it when we receive the notification.

As a first pass, there are a few things left out, like a few additional
sanity checks on the configuration, any kind of configurability, or the
ability to function in master only mode (simply starting, stopping, and
monitoring an mfs-master withou handling a metalogger, promotion, and
so on).