Handle cleanly shutdown and failed master nodes
When promoting to master, we have three possible sources of master data to
choose from; a metadata.mfs from a cleanly shutdown master, a
metadata.mfs.back from a crashed master, along with associated changelog.*.mfs
logs that need to be replayed, or a metadata_ml.mfs.back from the metalogger,
along with its changelog_ml.*.mfs logs that need to be replayed.

This change simply naively picks between these three in that order, which
is wrong for reasons described in a comment.  We need to either ensure that
we clean up so we only have the latest data available in one of those three
forms, or add a way to distinguish which has the most recent data.