Catch more errors, fix more permissions
Errors writing the configuration file should be caught, including the
write we do just to set up our parameters to be able to query the status
of the server.

Also, if we're a disconnected shadow master, check to see if we expect
to be connected.  If so, wait a few seconds for a connection (as we are
notified before the master is promoted), but if we still haven't
connected, return an error.  This will help shut down stuck shadow
masters rather than having them continually try and fail to connect.